+++ client/src/components/pages/Modeler/Modeler.jsx
@@ -4,7 +4,7 @@ import BpmnModeler from '../../content/BpmnModeler/BpmnModeler';
 import HeaderNavbar from '../../content/HeaderNavbar/HeaderNavbar';
 import ModelerSidebar from '../../content/ModelerSidebar/ModelerSidebar';
 import { getSsot } from '../../../api/ssot';
-import { getAllParametersForRobot } from '../../../api/variableRetrieval';
+import { getAllParametersForRobot } from '../../../api/parameterRetrieval';
 import { getAllAttributes } from '../../../api/attributeRetrieval';
 import { getAllRpaFunctionalities } from '../../../api/applicationAndTaskSelection';
 import { setRobotId } from '../../../utils/localSsot/ssot';
