+++ server/api/routes/robots/parameters/parameters.js
@@ -1,10 +1,10 @@
 const express = require('express');
-const variableController = require('../../../controllers/ssotVariableController');
+const parameterController = require('../../../controllers/ssotParameterController');
 
 const router = express.Router();
 
-router.put('/', variableController.updateMany);
-router.delete('/:robotId', variableController.deleteForActivities);
-router.get('/:robotId', variableController.retrieveParametersForRobot);
+router.put('/', parameterController.updateMany);
+router.delete('/:robotId', parameterController.deleteForActivities);
+router.get('/:robotId', parameterController.retrieveParametersForRobot);
 
 module.exports = router;
