+++ server/api/routes/robots/parameters/parameters.test.js
@@ -4,7 +4,7 @@ const mongoose = require('mongoose');
 const httpMocks = require('node-mocks-http');
 const dbHandler = require('../../../../utils/TestingUtils/TestDatabaseHandler');
 const dbLoader = require('../../../../utils/TestingUtils/databaseLoader');
-const ssotVariableController = require('../../../controllers/ssotVariableController');
+const ssotParameterController = require('../../../controllers/ssotParameterController');
 
 // eslint-disable-next-line no-unused-vars
 const rpaTaskModel = require('../../../models/rpaTaskModel');
@@ -49,7 +49,7 @@ describe('PUT /robots/parameters', () => {
     });
     const response = httpMocks.createResponse();
 
-    await ssotVariableController.updateMany(request, response);
+    await ssotParameterController.updateMany(request, response);
     expect(response.statusCode).toBe(200);
     const data = await response._getData();
     expect(data.modifiedCount).toBe(1);
@@ -78,7 +78,7 @@ describe('GET /robots/parameters/{robotId}', () => {
     });
     const response = httpMocks.createResponse();
 
-    await ssotVariableController.retrieveParametersForRobot(request, response);
+    await ssotParameterController.retrieveParametersForRobot(request, response);
     expect(response.statusCode).toBe(200);
     const data = await response._getData();
     expect(data.length).toBe(3);
@@ -112,7 +112,7 @@ describe('DELETE /robots/parameters/{robotId}', () => {
     });
     const response = httpMocks.createResponse();
 
-    await ssotVariableController.deleteForActivities(request, response);
+    await ssotParameterController.deleteForActivities(request, response);
 
     const foundParameters = await mongoose.model('parameter').find().exec();
     expect(foundParameters.length).toBe(1);
